Fixed error in trainer editor GUI reading AI flags from sheets; tried to address a Windows-only bug where a file deletion error occurs when the program initializes trainer text at project open by ensuring the folder exists before attempting to clear it

+++ src/com/turtleisaac/pokeditor/editors/trainers/gen4/TrainerTextEditor.java
@@ -86,7 +86,10 @@ public class TrainerTextEditor
     {
         try
         {
-            FileUtils.deleteDirectory(folder);
+            if (folder.exists())
+            {
+                FileUtils.deleteDirectory(folder);
+            }
         }
         catch(IOException e)
         {
